refactor(feishu): colocate participant provider
